feat: Add CRUD operations for User and Item models
This commit adds CRUD (Create, Read, Update, Delete) operations for the User and Item models in the app. It includes functions to get a user by ID or email, get a list of users with pagination, create a new user, get a list of items with pagination, and create a new item associated with a user.

These changes are necessary to enable the app to perform basic database operations on users and items, which will be used in upcoming features related to user authentication and item management.